Extract random creation of pubkeys for testing
